doc: Be more specific to supported CPU models for CPU templates
We are going to add a new Intel CPU (Sapphire Rapids). But the existing
CPU templates cannot be used on Sapphire Rapids. The reason is as
follows. Sapphire Rapids supports a wider CPUID range. To adapt the
existing templates to Sapphire Rapids, they have to mask CPUID bits off
on the new CPUID range. However, if such a CPU template is applied on
older CPUs that don't have the new CPUID range, Firecracker exits with
an error. This error is intentional because the CPU template feature has
the great power of tweaking very low layer stuff and fail-close is
preferable.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>