chore(test): Refine exception list for rdmsr consistency check
Some MSRs were not needed to be excepted and others had wrong comments.
Make the exception list minimal and leave a more specific comment for
each MSR. We might have to readd some of removed ones when upgrading
guest kernels, but let's leave more details for them in that case. I'm
the one who will update guest kernels to make a new release, so don't
worry :)

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>