test: Support all CPU templates in test_cpu_rdmsr
There was a long lasting TODO that validates all the CPU templates. For
CPU templates that do not touch any MSR, all the MSR values must be the
same as the case not applying any CPU template. Thus, we can remove MSR
baselines for them instead use ones for no CPU template.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>