test: only print uffd hint if uffd actually in use
If we fail to kill firecracker because it already die, we print a hint
about the uffd handler being potentially responsible. But this only
makes sense if a uffd handler was actually spawned, so restrict printing
of the hint to this case. While we're at it, also print the uffd logs in
this case, so that the truth of the hint can easily be verified (uffd
killed firecracker iff uffd panicked).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>