refactor(test): unify debug info dumping routines
We have two places where the test framework dumps a bunch of debug info:
When something fails via SSH, and when we fail to kill a microVM because
it's already dead. Let's unify both of these to dump the _same_ debug
information, by having both call _dump_debug_information (generalized
slightly so that it no longer hardcodes "SSH gone wrong" as the error
message, and extended to also print uffd logs).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>